refactor: delete SettingsBridge — AutotileEngine uses IAutotileSettings directly
Removes 853 lines of settings bridge code. AutotileEngine now follows
the exact same pattern as SnapEngine:
- autotileSettings() returns IAutotileSettings* via dynamic_cast from
  PlacementEngineBase::engineSettings()
- refreshConfigFromSettings() reads all settings into AutotileConfig
  on demand (called by daemon after any settings change)
- Write-back uses settingsWriteBackRequested(QVariantMap) signal
  (replaces syncAlgorithmToSettings + syncShortcutAdjustment)
- Serialization (window orders, pending restores) inlined into engine

IPlacementEngine gains refreshConfigFromSettings() as a virtual no-op,
replacing the old syncFromSettings(QObject*) + connectToSettings(QObject*).
Engines that cache config override it; engines that read on demand don't.

Both engines now have identical settings patterns:
  1. Daemon calls setEngineSettings(settingsQObject) once
  2. Engine casts to its specific interface at point of use
  3. Daemon connects settings change signals to retile externally
  4. Engine emits settingsWriteBackRequested for persistence

SettingsBridge test removed (tested deleted code). 130/130 tests pass.
