+++ src/daemon/enginefactory.cpp
@@ -20,6 +20,7 @@ EngineSet createEngines(PhosphorZones::LayoutRegistry* layoutManager, WindowTrac
     auto autotile =
         std::make_unique<AutotileEngine>(layoutManager, windowTracker, screenManager, algorithmRegistry, parent);
     autotile->setWindowRegistry(windowRegistry);
+    autotile->setEngineSettings(settings);
 
     // --- SnapEngine ---
     auto snap = std::make_unique<SnapEngine>(layoutManager, windowTracker, zoneDetector, vdm, parent);
