+++ tests/unit/CMakeLists.txt
@@ -287,10 +287,7 @@ target_compile_definitions(test_per_screen_config PRIVATE "PZ_SOURCE_DIR=\"${PRO
 pz_add_test(test_navigation_controller autotile/test_navigation_controller.cpp)
 
 # ═══════════════════════════════════════════════════════════════════════════════
-# autotile/ — Settings Bridge Tests
-# ═══════════════════════════════════════════════════════════════════════════════
-pz_add_test(test_settings_bridge autotile/test_settings_bridge.cpp)
-target_compile_definitions(test_settings_bridge PRIVATE "PZ_SOURCE_DIR=\"${PROJECT_SOURCE_DIR}\"")
+# autotile/ — Settings Bridge Tests (removed: SettingsBridge deleted, engine uses IAutotileSettings directly)
 
 # ═══════════════════════════════════════════════════════════════════════════════
 # ui/ — Zone Shader Item Tests — need Qt6::Quick + plasmazones_rendering
